redid NPC state diagram
NPC goes to idle by default. If it is not healed, it will prowl aimlessly. If it sees the player, it will walk towards them faster. If it is within range of the player, it will turn on bool attack and attack as long as they are within range. If it is healed, it will go through a healing animation, then to idle, and then to the following state.